4.21

+++ controllers/blogs.js
@@ -12,9 +12,7 @@ blogsRouter.get('/', async (req, res) => {
 
 blogsRouter.post('/', async (req, res) => {
   const body = req.body
-  console.log(body)
   const decodedToken = jwt.verify(req.token, process.env.SECRET)
-  console.log(decodedToken)
   if (!req.token || !decodedToken.id) {
     return res.status(401).json({ error: 'token missing or invalid' })
   }
@@ -40,8 +38,20 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req,res) => {
-  await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  const blog = await Blog.findById(req.params.id)
+  if (!blog){
+    return res.status(404).json({ error: 'no blog found' })
+  }
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing' })
+  }
+  if ( blog.user.toString() === decodedToken.id.toString() ){
+    await Blog.findByIdAndDelete(req.params.id)
+    res.status(204).end()
+  } else {
+    return res.status(401).json({ error: 'user not allowed' })
+  }
 })
 
 blogsRouter.put('/:id', async (req,res) => {
